Moved as much functionality as possible to helperFuncs.js, Article component was unnecessarily heavy
