+++ web-app/client/utils/helperFuncs.js
@@ -18,3 +18,31 @@ export function setLocalInteractions (newInteractionsArr) {
 export function getLastMemberOf (arr) {
   if (arr.length && arr.length > 0) return arr[arr.length - 1];
 }
+
+export function addInteractionToLocalStorage (articleId) {
+  const newInteraction = {
+    articleId: Number(articleId),
+    startTime: new Date(),
+    endTime: new Date()
+  };
+  if (localInteractionsExist()) {
+    const existingInteractions = getLocalInteractions();
+    setLocalInteractions(existingInteractions.concat(newInteraction));
+  } else {
+    setLocalInteractions([newInteraction]);
+  }
+}
+
+export function updateLastInteractionEndTime () {
+  if (localInteractionsExist()) {
+    const existingInteractions = getLocalInteractions();
+    const lastInteraction = getLastMemberOf(existingInteractions);
+    const updatedLastInteraction = Object.assign({}, lastInteraction, {
+      endTime: new Date()
+    });
+    const updatedInteractions = existingInteractions
+      .slice(0, existingInteractions.length - 1)
+      .concat(updatedLastInteraction);
+    setLocalInteractions(updatedInteractions);
+  }
+}
